* Removing symbolic links in experiments/data/finance (closes #275) * Adding PostgreSQL result files for all TPCH and financial queries. * Code formatting of Scala source files (closes #202). * Resolving problems with DB parser. All TPCH result files use '|' as delimiter and quotes around strings. * Minor tweaks to query_test.rb: the temporary file is deleted at exit, db_parser is extended to support semicolons in strings.
git-svn-id: file:///Users/ckoch/repos/assembla.com/boo@2554 e0e46c49-be69-4f5a-ad62-21024a331aea

Former-commit-id: 1ad0f8e8836a249d44ecc922a201e5515074f604

+++ test/scripts/postgres/finance/schemas.sql
@@ -4,19 +4,19 @@ CREATE SCHEMA "FINANCE_@@DATASET@@"
   AUTHORIZATION postgres;
 
 CREATE TABLE IF NOT EXISTS "FINANCE_@@DATASET@@".ASKS (
-  T         float(53), 
+  T         float, 
   ID        int, 
   BROKER_ID int, 
-  VOLUME    float(53), 
-  PRICE     float(53)
+  VOLUME    float, 
+  PRICE     float
 );
 
 CREATE TABLE IF NOT EXISTS "FINANCE_@@DATASET@@".BIDS (
-  T         float(53), 
+  T         float, 
   ID        int, 
   BROKER_ID int, 
-  VOLUME    float(53), 
-  PRICE     float(53)
+  VOLUME    float,
+  PRICE     float
 );
